backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.7
django-ckeditor==6.7.0
django-ckeditor-5==0.2.12
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-hosts==6.0
django-js-asset==2.2.0
django-model-utils==5.0.0
django-nine==0.2.7
django-notifications-hq==1.8.3
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.2
elevenlabs==1.54.0
executing==2.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
icecream==2.1.4
idna==3.8
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonfield==3.1.0
# msgpack==1.0.8
msgpack
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.2
pluggy==1.5.0
psycopg2-binary
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.2
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.6.4
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.0.3
paramiko==3.5.1
django-ckeditor
SpeechRecognition
pydub
openai==0.28.1
fitz
frontend
tools
python-docx
ffmpeg

